requirements.txt
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
async-generator==1.10
asyncio==3.4.3
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudinary==1.29.0
colorama==0.4.4
coverage==6.3.2
cryptography==36.0.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.62.0
flake8==4.0.1
h11==0.12.0
httpcore==0.14.7
httptools==0.2.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
mccabe==0.6.1
motor==2.3.1
mypy-extensions==0.4.3
orjson==3.6.7
outcome==1.1.0
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.0
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
pymongo==3.12.3
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
regex==2022.1.18
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.13.6
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
